filter buttons moved to Bar.js
